Hotfix Release: use system-provided cyrus-sasl/libsasl2 at runtime (#9)
* Use system-provided cyrus-sasl/libsasl2 at runtime

Instead of having multiple builds with and without `libsasl2` on Linux,
one of which cannot be redistributed within Python wheels, we use
`dlopen` to load `libsasl2` at runtime on Unix.
The SASL GSSAPI mechanism availability is thus checked at runtime.
Because of differences in soname ABI bumps between Debian-based and
RPM-based distros, the previous SASL builds did not work on Debian-based
systems. This is also solved in this change, by probing the various
known names for `libsasl2`.

* Compute version from tag or from vcpkg.json

* Run style checks on tag too

* Re-enable tests that were disabled in our CI

* Allow build metadata in nuget version